double anti-derivative :baguette_bread:
thanks to macros I was able to learn, for probably the third time,
how the newton's method thing works. for the rest of the integrals,
I just kinda cheesed it based on what I already know of derivatives.
if it works, is it stupid? the answer is yes. I have no idea what
I'm doing.